testes para repository

+++ src/main/java/com/ravel/votacaoapi/controller/VotacaoController.java
@@ -39,7 +39,7 @@ public class VotacaoController {
     }
 
     @GetMapping(value = "/contabilizar")
-    public Map<String, Long> contabilizarVotos(@RequestParam Long pautaId) {
+    public Map<String, Long> contabilizarVotos(@RequestParam int pautaId) {
         return service.contabilizarVotos(pautaId);
     }
